+++ CHANGELOG.md
@@ -29,7 +29,7 @@ If you like this project and find it useful, please consider giving it a star on
 
 <a href="https://www.buymeacoffee.com/luligugithub"><img src="https://matterbridge.io/assets/bmc-button.svg" alt="Buy me a coffee" width="120"></a>
 
-## [1.7.0] - 2026-08-29
+## [1.7.0] - Dev branch
 
 ### Breaking changes